refactor: Split DAGs in ETH and ETC
