+++ dags/ethereum_load_dag.py
@@ -1,0 +1,7 @@
+from __future__ import print_function
+
+import logging
+
+from airflow.models import Variable
+
+from ethereumetl.build_load_dag import build_load_dag
